% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{adl.dummy.checks}
\alias{adl.dummy.checks}
\title{Do consistent dummy checks for functions (GDTE/GDRF) that use an ADL model}
\usage{
adl.dummy.checks(
  x.vrbl,
  y.vrbl,
  d.x,
  d.y,
  inferences.x,
  inferences.y,
  the.coef,
  se.type,
  type = NULL
)
}
\arguments{
\item{x.vrbl}{a named vector of the x variables and corresponding lag orders in an ADL model}

\item{y.vrbl}{a named vector of the y variables and corresponding lag orders in an ADL model}

\item{d.x}{the order of differencing of the x variable in the ADL model}

\item{d.y}{the order of differencing of the y variable in the ADL model}

\item{inferences.x}{is the independent variable treated in levels or in differences?}

\item{inferences.y}{are the inferences for the dependent variable expected in levels or in differences?}

\item{the.coef}{the coefficient vector from the estimated ADL model}

\item{se.type}{the type of standard error calculated}

\item{type}{whether the effects are estimated in the context of a GDTE/GDRF}
}
\description{
Do consistent dummy checks for functions (GDTE/GDRF) that use an ADL model
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{internal}
