% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{adl.plot}
\alias{adl.plot}
\title{Evaluate (and possibly plot) the General Dynamic Response Function (GDRF) for an autoregressive distributed lag (ADL) model, assuming the underlying model is in levels (\code{d.x} = \code{d.y} = 0) and the user wants a marginal effect (the untransformed GDRF). (This is just a wrapper for \code{GDRF.adl.plot} with simplifying assumptions)}
\usage{
adl.plot(
  model = NULL,
  x.vrbl = NULL,
  y.vrbl = NULL,
  shock.history = "pulse",
  dM.level = 0.95,
  s.limit = 20,
  se.type = "const",
  return.data = FALSE,
  return.plot = TRUE,
  return.formulae = FALSE,
  ...
)
}
\arguments{
\item{model}{the \code{lm} model containing the ADL estimates}

\item{x.vrbl}{a named numeric vector in which the names correspond to an independent variable and its lags and the numbers correspond to the specific lag order of each variable}

\item{y.vrbl}{a named numeric vector in which the names correspond to lags of the dependent variable and the numbers correspond to the specific lag order of each variable. Can be \code{NULL} if the model has no lagged dependent variables}

\item{shock.history}{the desired shock history. \code{shock.history} determines the shock history (h) (which can be expressed as an integer) that will be applied to the independent variable. -1 represents a pulse (Impulse Response Function). 0 represents a step (Step Response Function). These can also be specified via \code{pulse} and \code{step}. For others, see Vande Kamp, Jordan, and Rajan. The default is \code{pulse}}

\item{dM.level}{a numeric significance level of the GDRF, calculated by the delta method. The default is 0.95}

\item{s.limit}{an integer for the number of periods to determine the GDRF (beginning at s = 0)}

\item{se.type}{a string for the type of standard error to extract from the model. The default is \code{const}, but any argument to \code{vcovHC} from the \code{sandwich} package is accepted}

\item{return.data}{logical to return the raw calculated GDRFs as a list element under \code{estimates}. The default is \code{FALSE}}

\item{return.plot}{logical to return the visualized GDRFs as a list element under \code{plot}. The default is \code{TRUE}}

\item{return.formulae}{logical to return the formulae for the GDRFs as a list element under \code{formulae} (for the GDRFs) and \code{binomials} (for the shock history). The default is \code{FALSE}}

\item{...}{other arguments to be passed to the call to plot}
}
\description{
Evaluate (and possibly plot) the General Dynamic Response Function (GDRF) for an autoregressive distributed lag (ADL) model, assuming the underlying model is in levels (\code{d.x} = \code{d.y} = 0) and the user wants a marginal effect (the untransformed GDRF). (This is just a wrapper for \code{GDRF.adl.plot} with simplifying assumptions)
}
\examples{

# ADL(1,1)
# Use the toy data to run an ADL. No argument is made this is well specified; it is just expository
model.toydata <- lm(y ~ l_1_y + x + l_1_x, data = toy.ts.interaction.data)

# Since this is in levels, we can quickly look at the adl.plot
#  Pulse effect of x
adl.plot(model = model.toydata,
	x.vrbl = c("x" = 0, "l_1_x" = 1),
	y.vrbl = c("l_1_y" = 1),
	shock.history = "pulse",
	s.limit = 20)

}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{ADL}
\keyword{plot}
