% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsinteractions.R
\docType{data}
\name{approval}
\alias{approval}
\title{Data on US Presidential Approval}
\format{
A data frame with 140 rows and 14 variables:
\describe{
 \item{APPROVE}{Presidential approval}
 \item{APPROVE_ECONOMY}{Presidential approval: economy}
 \item{APPROVE_FOREIGN}{Presidential approval: foreign affairs}
 \item{MIP_MACROECONOMICS}{Salience (Most Important Problem): economy}
 \item{MIP_FOREIGN}{Salience (Most Important Problem): foreign affairs}
 \item{PARTY_IN}{Macropartisanship (in-party)}
 \item{PARTY_OUT}{Macropartisanship (out-party)}
 \item{PRESIDENT}{Numeric indicator for president}
 \item{DIVIDEDGOV}{Dummy variable for divided government}
 \item{ELECTION}{Dummy variable for election years}
 \item{HONEYMOON}{Dummy variable for honeymoon period}
 \item{UMCSENT}{Consumer sentiment}
 \item{UNRATE}{Unemployment rate}
 \item{APPROVE_L1}{Lagged presidential approval}
}
}
\source{
\doi{10.1111/psq.12594}
}
\usage{
data(approval)
}
\description{
A dataset from: Cavari, Amnon. 2019. "Evaluating the President on Your Priorities: Issue Priorities, Policy Performance, and Presidential Approval, 1981–2016." Presidential Studies Quarterly 49(4): 798-826.
}
\keyword{datasets}
