% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{GDRF.dummy.checks}
\alias{GDRF.dummy.checks}
\title{Do consistent dummy checks for GDRF functions that might take fitted values}
\usage{
GDRF.dummy.checks(
  effect.type,
  prediction.values,
  baseline.y,
  baseline.y.se,
  shock.size,
  d.y,
  inferences.y
)
}
\arguments{
\item{effect.type}{whether to return marginal effects or fitted values. \code{marginal} returns the GDRF as a marginal effect. \code{fitted} returns the GDRF as a fitted value, relative to a baseline value of y. The default is \code{marginal}}

\item{prediction.values}{a named list of values for non-y variables in the model, used to calculate a steady-state baseline when \code{effect.type = "fitted"} and \code{d.y = 0} and \code{baseline.y} is not supplied. This allows for the calculation of model-based uncertainty. If any differenced variables are included in the model, they should be set to 0. Ignored when \code{d.y > 0}}

\item{baseline.y}{a user-supplied baseline value of y in levels. For \code{d.y = 0}, this overrides the steady-state calculation from \code{prediction.values} if provided. For \code{d.y > 0} with \code{inferences.y = "levels"}, this is required (otherwise it is just marginal effects). Only used when \code{effect.type = "fitted"}}

\item{baseline.y.se}{a user-supplied standard error for the baseline value of y (to suggest uncertainty around predictions). If supplied, this is added in quadrature to the standard errors of the GDRF estimates. Only used when \code{effect.type = "fitted"} and \code{inferences.y = "levels"}. The default is 0: in recognition that this is user-constructed uncertainty. Possible values would be the square root of the standard deviation of y (in levels)}

\item{shock.size}{the size of the shock to x in the units of x. Only used when \code{effect.type = "fitted"}; marginal effects are not scaled. Defaults to 1 (a marginal effect)}

\item{d.y}{the order of differencing of the y variable in the ADL model}

\item{inferences.y}{does the user want resulting inferences about the dependent variable in levels or in differences? (For y variables where \code{d.y} is 0, this is automatically levels.) The default is \code{levels}}
}
\description{
Do consistent dummy checks for GDRF functions that might take fitted values
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{internal}
