% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{get.value}
\alias{get.value}
\title{Find starting values for predicted values plots from the data in the model frame, if not supplied}
\usage{
get.value(var, prediction.values, model)
}
\arguments{
\item{var}{the variable to establish a prediction value for}

\item{prediction.values}{(possible) user-supplied list of values for variables}

\item{model}{the model containing the dataframe for mean estimation if values are not user-supplied (and warn the user if we're taking the mean)}
}
\description{
Find starting values for predicted values plots from the data in the model frame, if not supplied
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{internal}
