% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{mpoly.subber}
\alias{mpoly.subber}
\title{Replace characters that mpoly does not take with underscores}
\usage{
mpoly.subber(env = environment())
}
\arguments{
\item{env}{the environment with names to substitute. Defaults to the parent environment}
}
\description{
Replace characters that mpoly does not take with underscores
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{internal}
