% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsinteractions.R
\docType{data}
\name{toy.ts.interaction.data}
\alias{toy.ts.interaction.data}
\title{Simulated interactive time series data}
\format{
A data frame with 50 rows and 23 variables:
\describe{
 \item{time}{Indicator for time period}
 \item{x}{Contemporaneous x}
 \item{l.1.x}{First lag of x}
 \item{l.2.x}{Second lag of x}
 \item{l.3.x}{Third lag of x}
 \item{l.4.x}{Fourth lag of x}
 \item{l.5.x}{Fifth lag of x}
 \item{z}{Contemporaneous z}
 \item{l.1.z}{First lag of z}
 \item{l.2.z}{Second lag of z}
 \item{l.3.z}{Third lag of z}
 \item{l.4.z}{Fourth lag of z}
 \item{l.5.z}{Fifth lag of z}
 \item{y}{Contemporaneous y}
 \item{l.1.y}{First lag of y}
 \item{l.2.y}{Second lag of y}
 \item{l.3.y}{Third lag of y}
 \item{l.4.y}{Fourth lag of y}
 \item{l.5.y}{Fifth lag of y}
 \item{x.z}{Interaction of contemporaneous x and z}
 \item{x.l.1.z}{Interaction of contemporaneous x and lagged z}
 \item{z.l.1.x}{Interaction of lagged x and contemporaneous z}
 \item{l.1.x.l.1.z}{Interaction of lagged x and lagged z}
}
}
\usage{
data(toy.ts.interaction.data)
}
\description{
A simulated, well-behaved dataset of interactive time series data
}
\keyword{datasets}
