% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{what.to.return}
\alias{what.to.return}
\title{Consistently return the correct objects after a GDRF ADL/GECM}
\usage{
what.to.return(
  return.plot,
  return.formulae,
  return.data,
  plot.out,
  dat.out,
  the.final.formulae
)
}
\arguments{
\item{return.plot}{a TRUE/FALSE on whether the plot should be returned from the function}

\item{return.formulae}{a TRUE/FALSE on whether the data from the effect should be returned from the function}

\item{plot.out}{the created plot from the GDRF ADL/GECM}

\item{dat.out}{the created data from the GDRF ADL/GECM}

\item{the.final.formulae}{the created formulae from the GDRF ADL/GECM}
}
\description{
Consistently return the correct objects after a GDRF ADL/GECM
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{internal}
