% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{pulse.calculator}
\alias{pulse.calculator}
\title{Generate pulse effect formulae for a given autoregressive distributed lag (ADL) model}
\usage{
pulse.calculator(x.vrbl, y.vrbl = NULL, limit)
}
\arguments{
\item{x.vrbl}{a named numeric vector in which the names correspond to an independent variable and its lags and the numbers correspond to the specific lag order of each variable}

\item{y.vrbl}{a named numeric vector in which the names correspond to lags of the dependent variable and the numbers correspond to the specific lag order of each variable. Can be \code{NULL} if the model has no lagged dependent variables}

\item{limit}{an integer representing the number of periods after the initial shock (s) to calculate the Impulse Response Function}
}
\value{
a list of limit + 1 \code{mpoly} formulae containing the pulse effect formula in each period
}
\description{
Generate pulse effect formulae for a given autoregressive distributed lag (ADL) model
}
\details{
\code{pulse.calculator} does no calculation. It generates a list of \code{mpoly} formulae that contain variable names that represent the pulse effect in each period. The expectation is that these will be evaluated using coefficients from an object containing an ADL model with corresponding variables. Note: \code{mpoly} does not allow variable names with a .; variables passed to \code{pulse.calculator} should not include this character
}
\examples{
# ADL(1,1)
x.lags <- c("x" = 0, "l_1_x" = 1) # lags of x
y.lags <- c("l_1_y" = 1)
s <- 5
pulses <- pulse.calculator(x.vrbl = x.lags, y.vrbl = y.lags, limit = s)
pulses
# Will also handle finite dynamics
x.lags <- c("x" = 0, "l_1_x" = 1) # lags of x
finite.pulses <- pulse.calculator(x.vrbl = x.lags, limit = s)
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{utilities}
