% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\docType{data}
\name{toy.ts.interaction.data}
\alias{toy.ts.interaction.data}
\title{Simulated interactive time series data}
\format{
A data frame with 50 rows and 23 variables:
\describe{
 \item{time}{Indicator for time period}
 \item{x}{Contemporaneous x}
 \item{l_1_x}{First lag of x}
 \item{l_2_x}{Second lag of x}
 \item{l_3_x}{Third lag of x}
 \item{l_4_x}{Fourth lag of x}
 \item{l_5_x}{Fifth lag of x}
 \item{d_x}{First difference of x}
 \item{l_1_d_x}{First lag of first difference of x}
 \item{l_2_d_x}{Second lag of first difference of x}
 \item{l_3_d_x}{Third lag of first difference of x}
 \item{z}{Contemporaneous z}
 \item{l_1_z}{First lag of z}
 \item{l_2_z}{Second lag of z}
 \item{l_3_z}{Third lag of z}
 \item{l_4_z}{Fourth lag of z}
 \item{l_5_z}{Fifth lag of z}
 \item{y}{Contemporaneous y}
 \item{l_1_y}{First lag of y}
 \item{l_2_y}{Second lag of y}
 \item{l_3_y}{Third lag of y}
 \item{l_4_y}{Fourth lag of y}
 \item{l_5_y}{Fifth lag of y}
 \item{d_y}{First difference of y}
 \item{l_1_d_y}{First lag of first difference of y}
 \item{l_2_d_y}{Second lag of first difference of y}
 \item{d_2_y}{Second difference of y}
 \item{l_1_d_2_y}{First lag of second difference of y}
 \item{x_z}{Interaction of contemporaneous x and z}
 \item{x_l_1_z}{Interaction of contemporaneous x and lagged z}
 \item{z_l_1_x}{Interaction of lagged x and contemporaneous z}
 \item{l_1_x_l_1_z}{Interaction of lagged x and lagged z}
}
}
\usage{
data(toy.ts.interaction.data)
}
\description{
A simulated, well-behaved dataset of interactive time series data
}
\keyword{datasets}
