% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseffects.R
\name{what.to.return}
\alias{what.to.return}
\title{Do consistent dummy checks for functions (GDTE/GDRF) that use a GECM model}
\usage{
what.to.return(
  return.plot,
  return.formulae,
  return.data,
  plot.out,
  dat.out,
  the.final.formulae
)
}
\arguments{
\item{return.plot}{a TRUE/FALSE on whether the plot should be returned from the function}

\item{return.formulae}{a TRUE/FALSE on whether the data from the effect should be returned from the function}
}
\description{
Do consistent dummy checks for functions (GDTE/GDRF) that use a GECM model
}
\author{
Soren Jordan, Garrett N. Vande Kamp, and Reshi Rajan
}
\keyword{internal}
